Update libcoveocds.libcore.tools.get_file_type and add tests for it
While copying tests from COVE, I found minor problems. It turns out the old version can accept strings of file names or Django file objects - but we can't test the latter here properly and we don't actually use that type of input at the moment.

So for now, updated function to work well with strings only, added tests that are like the ones in COVE but changed to strings.